+++ src/python_example/object_detection/yolov11m/yolov11m_async_cpp_postprocess.py
@@ -16,6 +16,13 @@ for _path in [str(_v3_dir), str(_module_dir)]:
     if _path not in sys.path:
         sys.path.insert(0, _path)
 
+
+import os
+if os.name == 'nt':
+    _dxrt_dir = os.environ.get('DXRT_DIR')
+    if _dxrt_dir:
+        os.add_dll_directory(os.path.join(_dxrt_dir, 'bin'))
+
 from dx_postprocess import YOLOv8PostProcess
 from dx_engine import InferenceOption
 from common.utility import convert_cpp_detections